at the bottom of page I want 2 links to privacy (for privacy statement) and disclaimer.
please use this text for  privacy, change it, make it simpler and for datasight.

Finidy B.V. is a service organisation. Finidy B.V. respects and protects the privacy of all users of the websites of Finidy B.V. Finidy B.V. makes every reasonable effort to ensure that all personal infor-mation provided to it will always be treated confidentially and will only be used for the purposes described in this Privacy Statement.

General information
Finidy B.V. is interested in information about which websites are most frequently visited and which services offered by Finidy B.V. are most in demand. Finidy B.V. collects information for this purpose by recording IP data and placing cookies. This information allows Finidy B.V. to adapt its websites and services to more effectively match user requirements. In order to protect website users' right to privacy, all information collected in these ways is general and anonymous. A user cannot be iden-tified based on this information. Cookies are also used to remember passwords, user names and preferences. You can set your browser to refuse cookies. However, this may have a negative im-pact on your website experience.

Links
The websites of Finidy B.V. contain links to websites of third parties. Finidy B.V. is not responsible for compliance with privacy legislation on the part of these third parties.

Supply of data to third parties
Finidy B.V. may be requested by the government or regulatory bodies to supply (personal) data as part of an investigation. Finidy B.V. reserves the right to provide processed (personal) data in re-sponse to a request of this nature.

Changes to the Privacy Policy
If Finidy B.V. decides to change its Privacy Policy, any such changes will be incorporated directly in the Privacy Policy presented here.

Questions and requests
If you have any questions or requests relating to Finidy B.V.'s Privacy Policy, please send them by email to info@finidy.nl

This Privacy Policy is governed by the laws of the Netherlands. Any disputes shall be submitted to the competent court in The Hague

please use this text for  disclaimer, change it, make it simpler and for datasight.

Finidy B.V. makes every possible effort to ensure that the data on its website is accurate and up-to-date. However, inaccuracies and omissions cannot be ruled out. Finidy B.V. reserves the right to change the content of the information on the website.

Finidy B.V. does not guarantee faultless operation or uninterrupted availability of the website or the services it offers electronically. Finidy B.V. is neither liable for damage arising from inaccuracies or omissions in the information offered, nor for damage resulting from problems caused by, or in-herent to, the distribution of information via the Internet, such as breakdowns or interruptions, or errors or delays in the provision of information by Finidy B.V., or by you to Finidy B.V., via a Finidy B.V. website or some other electronic means. Finidy B.V. cannot guarantee that the hyperlinks mentioned on the website will still be correct when you view them. Finidy B.V. accepts no respon-sibility for the content of websites that is referenced via a hyperlink or otherwise.

The General Terms and Conditions of Finidy B.V. apply to all agreements with Finidy B.V. Finidy B.V. reserves the right to change these current provisions. These General Terms and Conditions can be requested from Finidy B.V. by sending an email to info@finidy.nl. Are there differences between the General Terms and Conditions on paper and those on the website? If so, the version filed at the Chamber of Commerce is binding. Finidy B.V. does not guarantee that emails or other electron-ic messages sent to Finidy B.V. will be received and processed on time. Finidy B.V. does not accept any liability for the consequences of not receiving or processing them, or not receiving or pro-cessing them in good time.

Valid from 01/04/2024