I want to hide the publicaties section for now, as I don't have a publiciation yet.
The contact form i like to make an addition.on the outside of the form i like on the left side a curved arrow, that goes inside the the form to the "uw vraag of projectbeschrijving" that indicates a user that he can also contact us for a cup of (virtual) coffee, with a coffee emotion. Trye to be creative here.