i added my fav icons in the public directory, please implement them correctly in the website
in the public dir there's also id.svg, use that as the logo.
in the public dir theres als geometrical_bg_tom.png use that as the background image in the home section.  please use this image only as an overlag on the existing background color and use a larage opacity, needs to be slightly visible.